requirements.txt
django
ujson
djangorestframework
django-rq
ipython